tags and height fix
added generic red/green enemy/friend tag prefabs that are dynamically instantiated along with the virtual humans.

Fixed height of spawnPoint boxes to be on ground plane.

Added LookAt() code to virtual humans and tags to always face the main camera.

Adjusted layer settings of virtual humans to be visible on HL.

Added toggle switch on Spawner.cs to control visibility of virtual humans on HoloLens (independent from tag visibility).

In order for this to match the main Husis scene, we just need to adjust the spawn point positions in the other unity project to match the ones here (otherwise the height will be slightly off).